requirements.txt
Flask
Flask-Cors
moviepy
openai
pdfminer.six
python-dotenv
requests
# waitress
# Werkzeug